Unify similar tests in wizard budgets spec
We are unifying the test "Create budget - Knapsack voting (default)" with
"A new budget is always created in draft mode" because they are almost the same.
The only thing we changed from the original test is that instead of using:
> click_link "Go back to edit budget"
we are using:
> within("#side_menu") { click_link "Participatory budgets" }
> within("tr", text: "M30 - Summer campaign") { click_link "Edit" }

This way, we still test the "Edit" link, and the "Go back…" is already tested in
other specs.

On the other hand, we also merged the test "update budget" with "submit the
form with errors and then without errors". Just like in the previous case, there
were two ways to access the edit page, so we removed the one that is already
tested in other specs.