Correctly align radio buttons with their labels
We were using a `height: $line-height` property for this task. One of
the disadvantages of this approach is that things don't look so great
when the label expands over more than one line.

Back when we added that property, browser support for flex layouts
wasn't that great. Now there's universal support for it, so we can use
it instead.

Note that, right now, we're defining similar rules in many elements
instead of defining a general rule for every radio button with a label.
The reason is that in some cases we use `display: flex` and in other
places we use `display: inline-flex`. We might unify it in the future.